Removed bloat with xhr
Put the code to get json from an api in to the util file, changed the other instances where I got json data to use this function. There are a lot of scripts, I need to invest in a package manager because ES6 import is yet to be implemented in chromium and firefox.